requirements/base.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-environ==0.8.1
django-import-export==2.8.0
django-model-utils==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
Pillow==9.1.1
psycopg2==2.8.6
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0
xlrd==2.0.1
xlwt==1.3.0
